requirements.txt
hidapi>=0.11.0